ajodwyer/: Rename rcu_obj_base.hpp to simply rcu.hpp.
No intended functional change.
